Share script file scan concurrency utility
